Extract legacy tests from `SchemaExtenderTest`
