+++ tests/Utils/SchemaExtenderTest.php
@@ -1394,152 +1394,6 @@ class SchemaExtenderTest extends TestCase
     }
 
     /**
-     * @see it('does not allow replacing a custom directive')
-     */
-    public function testDoesNotAllowReplacingACustomDirective(): void
-    {
-        $extendedSchema = $this->extendTestSchema('
-          directive @meow(if: Boolean!) on FIELD | FRAGMENT_SPREAD
-        ');
-
-        $replacementAST = Parser::parse('
-            directive @meow(if: Boolean!) on FIELD | QUERY
-        ');
-
-        try {
-            SchemaExtender::extend($extendedSchema, $replacementAST);
-            self::fail();
-        } catch (Error $error) {
-            self::assertEquals('Directive "meow" already exists in the schema. It cannot be redefined.', $error->getMessage());
-        }
-    }
-
-    /**
-     * @see it('does not allow replacing an existing type')
-     */
-    public function testDoesNotAllowReplacingAnExistingType(): void
-    {
-        $existingTypeError = static function ($type): string {
-            return 'Type "' . $type . '" already exists in the schema. It cannot also be defined in this type definition.';
-        };
-
-        $typeSDL = '
-            type Bar
-        ';
-
-        try {
-            $this->extendTestSchema($typeSDL);
-            self::fail();
-        } catch (Error $error) {
-            self::assertEquals($existingTypeError('Bar'), $error->getMessage());
-        }
-
-        $scalarSDL = '
-          scalar SomeScalar
-        ';
-
-        try {
-            $this->extendTestSchema($scalarSDL);
-            self::fail();
-        } catch (Error $error) {
-            self::assertEquals($existingTypeError('SomeScalar'), $error->getMessage());
-        }
-
-        $interfaceSDL = '
-          interface SomeInterface
-        ';
-
-        try {
-            $this->extendTestSchema($interfaceSDL);
-            self::fail();
-        } catch (Error $error) {
-            self::assertEquals($existingTypeError('SomeInterface'), $error->getMessage());
-        }
-
-        $enumSDL = '
-          enum SomeEnum
-        ';
-
-        try {
-            $this->extendTestSchema($enumSDL);
-            self::fail();
-        } catch (Error $error) {
-            self::assertEquals($existingTypeError('SomeEnum'), $error->getMessage());
-        }
-
-        $unionSDL = '
-          union SomeUnion
-        ';
-
-        try {
-            $this->extendTestSchema($unionSDL);
-            self::fail();
-        } catch (Error $error) {
-            self::assertEquals($existingTypeError('SomeUnion'), $error->getMessage());
-        }
-
-        $inputSDL = '
-          input SomeInput
-        ';
-
-        try {
-            $this->extendTestSchema($inputSDL);
-            self::fail();
-        } catch (Error $error) {
-            self::assertEquals($existingTypeError('SomeInput'), $error->getMessage());
-        }
-    }
-
-    /**
-     * @see it('does not allow replacing an existing field')
-     */
-    public function testDoesNotAllowReplacingAnExistingField(): void
-    {
-        $existingFieldError = static function (string $type, string $field): string {
-            return 'Field "' . $type . '.' . $field . '" already exists in the schema. It cannot also be defined in this type extension.';
-        };
-
-        $typeSDL = '
-          extend type Bar {
-            foo: Foo
-          }
-        ';
-
-        try {
-            $this->extendTestSchema($typeSDL);
-            self::fail();
-        } catch (Error $error) {
-            self::assertEquals($existingFieldError('Bar', 'foo'), $error->getMessage());
-        }
-
-        $interfaceSDL = '
-          extend interface SomeInterface {
-            some: Foo
-          }
-        ';
-
-        try {
-            $this->extendTestSchema($interfaceSDL);
-            self::fail();
-        } catch (Error  $error) {
-            self::assertEquals($existingFieldError('SomeInterface', 'some'), $error->getMessage());
-        }
-
-        $inputSDL = '
-          extend input SomeInput {
-            fooArg: String
-          }
-        ';
-
-        try {
-            $this->extendTestSchema($inputSDL);
-            self::fail();
-        } catch (Error $error) {
-            self::assertEquals($existingFieldError('SomeInput', 'fooArg'), $error->getMessage());
-        }
-    }
-
-    /**
      * @see it('does not allow replacing an existing enum value')
      */
     public function testDoesNotAllowReplacingAnExistingEnumValue(): void
@@ -1555,149 +1409,6 @@ class SchemaExtenderTest extends TestCase
             self::fail();
         } catch (Error $error) {
             self::assertEquals('Enum value "SomeEnum.ONE" already exists in the schema. It cannot also be defined in this type extension.', $error->getMessage());
-        }
-    }
-
-    /**
-     * @see it('does not allow referencing an unknown type')
-     */
-    public function testDoesNotAllowReferencingAnUnknownType(): void
-    {
-        $unknownTypeError = 'Unknown type: "Quix". Ensure that this type exists either in the original schema, or is added in a type definition.';
-
-        $typeSDL = '
-          extend type Bar {
-            quix: Quix
-          }
-        ';
-
-        try {
-            $this->extendTestSchema($typeSDL);
-            self::fail();
-        } catch (Error $error) {
-            self::assertEquals($unknownTypeError, $error->getMessage());
-        }
-
-        $interfaceSDL = '
-          extend interface SomeInterface {
-            quix: Quix
-          }
-        ';
-
-        try {
-            $this->extendTestSchema($interfaceSDL);
-            self::fail();
-        } catch (Error $error) {
-            self::assertEquals($unknownTypeError, $error->getMessage());
-        }
-
-        $unionSDL = '
-          extend union SomeUnion = Quix
-        ';
-
-        try {
-            $this->extendTestSchema($unionSDL);
-            self::fail();
-        } catch (Error $error) {
-            self::assertEquals($unknownTypeError, $error->getMessage());
-        }
-
-        $inputSDL = '
-          extend input SomeInput {
-            quix: Quix
-          }
-        ';
-
-        try {
-            $this->extendTestSchema($inputSDL);
-            self::fail();
-        } catch (Error $error) {
-            self::assertEquals($unknownTypeError, $error->getMessage());
-        }
-    }
-
-    /**
-     * @see it('does not allow extending an unknown type')
-     */
-    public function testDoesNotAllowExtendingAnUnknownType(): void
-    {
-        $sdls = [
-            'extend scalar UnknownType @foo',
-            'extend type UnknownType @foo',
-            'extend interface UnknownType @foo',
-            'extend enum UnknownType @foo',
-            'extend union UnknownType @foo',
-            'extend input UnknownType @foo',
-        ];
-
-        foreach ($sdls as $sdl) {
-            try {
-                $this->extendTestSchema($sdl);
-                self::fail();
-            } catch (Error $error) {
-                self::assertEquals('Cannot extend type "UnknownType" because it does not exist in the existing schema.', $error->getMessage());
-            }
-        }
-    }
-
-    /**
-     * @see it('does not allow extending a mismatch type')
-     */
-    public function testDoesNotAllowExtendingAMismatchType(): void
-    {
-        $typeSDL = '
-          extend type SomeInterface @foo
-        ';
-
-        try {
-            $this->extendTestSchema($typeSDL);
-            self::fail();
-        } catch (Error $error) {
-            self::assertEquals('Cannot extend non-object type "SomeInterface".', $error->getMessage());
-        }
-
-        $interfaceSDL = '
-          extend interface Foo @foo
-        ';
-
-        try {
-            $this->extendTestSchema($interfaceSDL);
-            self::fail();
-        } catch (Error $error) {
-            self::assertEquals('Cannot extend non-interface type "Foo".', $error->getMessage());
-        }
-
-        $enumSDL = '
-          extend enum Foo @foo
-        ';
-
-        try {
-            $this->extendTestSchema($enumSDL);
-            self::fail();
-        } catch (Error $error) {
-            self::assertEquals('Cannot extend non-enum type "Foo".', $error->getMessage());
-        }
-
-        $unionSDL = '
-          extend union Foo @foo
-        ';
-
-        try {
-            $this->extendTestSchema($unionSDL);
-            self::fail();
-        } catch (Error $error) {
-            self::assertEquals('Cannot extend non-union type "Foo".', $error->getMessage());
-        }
-
-        $inputSDL = '
-          extend input Foo @foo
-        ';
-
-        try {
-            $this->extendTestSchema($inputSDL);
-            self::fail();
-        } catch (Error $error) {
-            self::assertEquals('Cannot extend non-input object type "Foo".', $error->getMessage());
         }
     }
 
@@ -1853,82 +1564,6 @@ class SchemaExtenderTest extends TestCase
                 }, $nodes)
             )
         );
-    }
-
-    /**
-     * @see it('does not allow redefining an existing root type')
-     */
-    public function testDoesNotAllowRedefiningAnExistingRootType(): void
-    {
-        $sdl = '
-            extend schema {
-                query: SomeType
-            }
-
-            type SomeType {
-                seeSomething: String
-            }
-        ';
-
-        try {
-            $this->extendTestSchema($sdl);
-            self::fail();
-        } catch (Error $error) {
-            self::assertEquals('Must provide only one query type in schema.', $error->getMessage());
-        }
-    }
-
-    /**
-     * @see it('does not allow defining a root operation type twice')
-     */
-    public function testDoesNotAllowDefiningARootOperationTypeTwice(): void
-    {
-        $sdl = '
-            extend schema {
-              mutation: Mutation
-            }
-            extend schema {
-              mutation: Mutation
-            }
-            type Mutation {
-              doSomething: String
-            }
-        ';
-
-        try {
-            $this->extendTestSchema($sdl);
-            self::fail();
-        } catch (Error $error) {
-            self::assertEquals('Must provide only one mutation type in schema.', $error->getMessage());
-        }
-    }
-
-    /**
-     * @see it('does not allow defining a root operation type with different types')
-     */
-    public function testDoesNotAllowDefiningARootOperationTypeWithDifferentTypes(): void
-    {
-        $sdl = '
-            extend schema {
-              mutation: Mutation
-            }
-            extend schema {
-              mutation: SomethingElse
-            }
-            type Mutation {
-              doSomething: String
-            }
-            type SomethingElse {
-              doSomethingElse: String
-            }
-        ';
-
-        try {
-            $this->extendTestSchema($sdl);
-            self::fail();
-        } catch (Error $error) {
-            self::assertEquals('Must provide only one mutation type in schema.', $error->getMessage());
-        }
     }
 
     /**
